Improved QDaqVector definition, added pop() function.
Adding the pop() required major rewrite of the underlying buffer object
in order to take into account the possibility that a full circular
buffer can become non-full after calling pop(). The buffer definition
was moved to math_util.h.
